Create Loader.Config Module
Status Quo: Configuration for for the loader does not have a central
definition.

Problem: This means that it is difficult to know exactly what options
are valid configuration options as well as what type of data each option
should hold. This also mean that validating configuration options must
be done all over the place instead of in one central location. This is a
introduces the possibility for one function to validate an option in one
way and another function to validate the option in a completely
different, possible incompatible way.

Solution: Create a Config module that provides a central definition for
the shape and type of the configuration object. The module exports
funtions for the creation and modification of a configuration object.
Its also the perfect place for any option validation to take place.