Create Loader.Output Module
Status Quo: the function that builds the output string for the loader in
currently found in the Handlebars.js file.

Problem: This is fine, however, we want to be able to have multiple
output formats. This could potentially clutter the Handlebars.js file.

Solution: Create a module to hold the different output formatting
functions.