Move Environment Setup Function To Handlebars Module
Status Quo: The function "setup_handlebars_environment" is found in
the main index.js file. It takes care of setting up our custom
handlebars runtime.

Problem: The name does not suit the purpose of the function. It would
make more sense for the file to be located in the Handlebars module
since it's role is primarily "handlebars" specific.

Solution: Rename the function to "setup_runtime" and move it to the
Handlebars module.