Setup Input Preprcessor Mechanism
Status Quo: The input source is compiled in whatever format in recieves.
The assumption is made that the input source is in a format that we
expect, with no way of accepting different types of input.

Problem: If the input enters our loader in an unexpected format, the
loader will choke. We want to have the flexibility of accepting input
from different loaders.

Solution: Setup a subsystem for processing different input format.