Setup Ouput Format Subsystem
Status Quo: We only have a single output format and no way for the user
to select b/w output formats.

Problem: We want to have the flexibility of configuring different ouput
formats that can be passed to different loaders.

Solution: Setup a configuration option for selecting the output format.
Create a function to select b/w different ouput formats. Update output
function to use take and produce a loader object or an error to
facilitate composability.