Move Option Description From Table To Own Subsection
Status Quo: In the README.md, the description for each option is kept in
a table along with the option name, type and default value.

Problem: This creats a really wide table that does not fit on the screen
in some cases and results in a squashed description field that is hard
to read.

Solution: Move the description from the table to it's own subsection
with a heading matching the corresponding option. Link each option name
in the table to it's corresponding subsection.