Remove "exclude" Config Option
Status Quo: The original handlebars-loader has an "exclude" config
option that, according to the original documentation, is used to
"exclude paths from resolving".

Problem: There is no immediately aparent use for this option. It could
be something that was added in response to some issue that a user had.
While this may be the case, we should not expose something on the API
with no apparent use.

Solution: Remove the "exclude" config option.